После нажатия на инлайн кнопку инлайн клавиатура пропадает и повторно нажать не получится
* Переписал функцию сохранения отзывов
* Создание инлайн клавиатуры вынес в файл keyboards
* Создал функцию set_route для установки состояния. Вынес её в functions
