+++ core/culture/management/commands/handlers/route.py
@@ -19,11 +19,11 @@ from ..config import logger, MAXIMUM_DURATION_VOICE_MESSAGE
 from ..crud import (get_all_exhibits_by_route, get_exhibit, get_route_by_id,
                     get_routes_id, save_review)
 from ..exceptions import FeedbackError
-from ..functions import (get_exhibit_from_state, get_id_from_state,
-                         get_route_from_state,
-                         speech_to_text_conversion)
-from ..keyboards import (KEYBOARD_YES_NO, make_row_keyboard,
-                         make_vertical_keyboard)
+from ..functions import (get_exhibit_from_state,
+                         get_id_from_state, get_route_from_state,
+                         set_route, speech_to_text_conversion)
+from ..keyboards import (KEYBOARD_YES_NO, keyboard_for_send_review,
+                         make_row_keyboard, make_vertical_keyboard)
 from ..utils import Route, Block
 from ..validators import rewiew_validator
 
@@ -250,32 +250,12 @@ async def exhibit(message: Message, state: FSMContext) -> None:
         await state.set_state(Route.review)
 
 
-def keyboard_for_send_review():
-    buttons = [
-        [
-            types.InlineKeyboardButton(
-                text='Оставить отзыв', callback_data='send_review'
-            ),
-            types.InlineKeyboardButton(
-                text='Без отзыва', callback_data='dont_send_review'
-            )
-        ]
-    ]
-    keyboard = types.InlineKeyboardMarkup(inline_keyboard=buttons)
-    return keyboard
-
-
 @route_router.message(Route.review, F.text | F.voice)
 async def review(message: Message, state: FSMContext) -> None:
     '''Получения отзыва'''
     global target
     target = True
     print(await state.get_data())
-    # builder = InlineKeyboardBuilder()
-    # builder.add(types.InlineKeyboardButton(
-    #     text='Оставить отзыв',
-    #     callback_data='send_review'
-    # ))
     answer = ''
     if message.voice:
         if message.voice.duration <= MAXIMUM_DURATION_VOICE_MESSAGE:
@@ -306,40 +286,16 @@ async def review(message: Message, state: FSMContext) -> None:
         await save_review(text, state)
         answer = ms.SUCCESSFUL_MESSAGE
         await message.answer(text=answer)
-
-        exhibit = await get_exhibit_from_state(state)
-
-        route_id, exhibit_number = await get_id_from_state(state)
-        exhibit_number += 1
-        await state.update_data(exhibit_number=exhibit_number)
-        route = await get_route_by_id(route_id)
-        if exhibit_number == len(await get_all_exhibits_by_route(route)):
-            await message.answer(
-                'Конец маршрута',
-                reply_markup=make_row_keyboard(['Конец']),
-            )
-            await state.set_state(Route.quiz)
-        else:
-            if exhibit.transfer_message != '':
-                await message.answer(
-                    f"{exhibit.transfer_message}",
-                )
-            await message.answer(
-                'Нас ждут длительные переходы',
-                reply_markup=make_row_keyboard(['Отлично идем дальше']),
-            )
-            exhibit = await get_exhibit(route_id, exhibit_number)
-            await state.update_data(exhibit_obj=exhibit)
-            await state.set_state(Route.transition)
+        await set_route(state, message)
     else:
         await message.answer(text=f'{answer}\nПопробуйте снова.',
                              reply_markup=keyboard_for_send_review())
-        # await state.set_state(Route.review)
 
 
 @route_router.callback_query(F.data == 'send_review')
 async def resend_review(callback: types.CallbackQuery, state: FSMContext):
     await callback.answer()
+    await callback.message.edit_reply_markup()
     await callback.message.answer('Напишите ваше мнение')
     await state.set_state(Route.review)
 
@@ -347,30 +303,8 @@ async def resend_review(callback: types.CallbackQuery, state: FSMContext):
 @route_router.callback_query(F.data == 'dont_send_review')
 async def skip_send_review(callback: types.CallbackQuery, state: FSMContext):
     await callback.answer()
-    exhibit = await get_exhibit_from_state(state)
-
-    route_id, exhibit_number = await get_id_from_state(state)
-    exhibit_number += 1
-    await state.update_data(exhibit_number=exhibit_number)
-    route = await get_route_by_id(route_id)
-    if exhibit_number == len(await get_all_exhibits_by_route(route)):
-        await callback.message.answer(
-            'Конец маршрута',
-            reply_markup=make_row_keyboard(['Конец']),
-        )
-        await state.set_state(Route.quiz)
-    else:
-        if exhibit.transfer_message != '':
-            await callback.message.answer(
-                f"{exhibit.transfer_message}",
-            )
-        await callback.message.answer(
-            'Нас ждут длительные переходы',
-            reply_markup=make_row_keyboard(['Отлично идем дальше']),
-        )
-        exhibit = await get_exhibit(route_id, exhibit_number)
-        await state.update_data(exhibit_obj=exhibit)
-        await state.set_state(Route.transition)
+    await callback.message.edit_reply_markup()
+    await set_route(state, callback.message)
 
 
 # этот код будет работать только если ботом пользуется только один человек,
